refactor(ui): extract AppShell and make Map a pure renderer
Extracts layout, widgets, responsive logic and URL-hash handling from
StandaloneApp into a new shared AppShell component, and strips all
data-loading out of Map.svelte. Standalone now owns its VectorSource
and the standalone-only pitch layer, composing AppShell with
single-backend providers.

Map's overlay hover hit-test switches from direct layer references to a
`layer.get('kind') === 'overlay'` filter so future overlays (standalone
pitches, hub overlays) can participate without Map knowing about them.

No user-visible behaviour change — unblocks the hub-ui-parity epic by
making AppShell reusable for the hub.

Refs #144
